initial migration and rename of ButtonComponent to Beta::Button
