+++ app/components/primer/dropdown.rb
@@ -4,7 +4,7 @@ module Primer
   # `Dropdown` is a lightweight context menu for housing navigation and actions.
   # They're great for instances where you don't need the full power (and code) of the SelectMenu.
   class Dropdown < Primer::Component
-    # Required trigger for the dropdown. Has the same arguments as <%= link_to_component(Primer::ButtonComponent) %>,
+    # Required trigger for the dropdown. Has the same arguments as <%= link_to_component(Primer::Beta::Button) %>,
     # but it is locked as a `summary` tag.
     renders_one :button, lambda { |**system_arguments|
       @button_arguments = system_arguments
