+++ app/components/primer/subhead_component.rb
@@ -5,7 +5,7 @@ module Primer
   #
   # - Optionally set the `:description` slot to render a short description and the `:actions` slot for a related action.
   # - Use a succinct, one-line description for the `:description` slot. For longer descriptions, omit the description slot and render a paragraph below the `Subhead`.
-  # - Use the actions slot to render a related action to the right of the heading. Use <%= link_to_component(Primer::ButtonComponent) %> or <%= link_to_component(Primer::LinkComponent) %>.
+  # - Use the actions slot to render a related action to the right of the heading. Use <%= link_to_component(Primer::Beta::Button) %> or <%= link_to_component(Primer::LinkComponent) %>.
   #
   # @accessibility
   #   The `:heading` slot defaults to rendering a `<div>`. Update the tag to a heading element with the appropriate level to improve page navigation for assistive technologies.
@@ -102,7 +102,7 @@ module Primer
     #     <% end %>
     #     <% component.actions do %>
     #       <%= render(
-    #         Primer::ButtonComponent.new(
+    #         Primer::Beta::Button.new(
     #           tag: :a, href: "http://www.google.com", scheme: :danger
     #         )
     #       ) { "Action" } %>
