+++ lib/primer/forms/submit_button.html.erb
@@ -1,4 +1,4 @@
-<%= render(Primer::ButtonComponent.new(**input_arguments)) do |c| %>
+<%= render(Primer::Beta::Button.new(**input_arguments)) do |c| %>
   <% @input.block.call(c) if @input.block %>
   <%= @input.label %>
 <% end %>
