+++ lib/primer/view_components/linters/button_component_migration_counter.rb
@@ -18,9 +18,9 @@ module ERBLint
       # CloseButton component has preference when this class is seen in conjunction with `btn`.
       DISALLOWED_CLASSES = %w[close-button].freeze
       CLASSES = %w[btn btn-link].freeze
-      MESSAGE = "We are migrating buttons to use [Primer::ButtonComponent](https://primer.style/view-components/components/button), please try to use that instead of raw HTML."
+      MESSAGE = "We are migrating buttons to use [Primer::Beta::Button](https://primer.style/view-components/components/button), please try to use that instead of raw HTML."
       ARGUMENT_MAPPER = ArgumentMappers::Button
-      COMPONENT = "Primer::ButtonComponent"
+      COMPONENT = "Primer::Beta::Button"
     end
   end
 end
