+++ static/statuses.json
@@ -22,6 +22,7 @@
   "Primer::Beta::BorderBox::Header": "alpha",
   "Primer::Beta::Breadcrumbs": "beta",
   "Primer::Beta::Breadcrumbs::Item": "alpha",
+  "Primer::Beta::Button": "beta",
   "Primer::Beta::Flash": "beta",
   "Primer::Beta::Text": "beta",
   "Primer::Beta::Truncate": "beta",
@@ -29,7 +30,7 @@
   "Primer::BlankslateComponent": "deprecated",
   "Primer::BorderBoxComponent": "deprecated",
   "Primer::BoxComponent": "stable",
-  "Primer::ButtonComponent": "beta",
+  "Primer::ButtonComponent": "deprecated",
   "Primer::ButtonGroup": "beta",
   "Primer::ClipboardCopy": "beta",
   "Primer::CloseButton": "beta",
